==> HAPIEE/6_Physical_measures/4_Obesity.Rmd <==
---
title: "Obesity"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/obesityDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/obesityDS.R')
```

# Data process


## Poland

### Wave 1

#### Study-specific variable description



| **Name** |`bmi`|
|-|-|
| **Label** |`Body Mass Index calculated as weight (kg)/height(m)^2`|
| **Table name** |`hapiee_pol_w1_hd`|
| **Type of variable** |`continuous`|
| **Missings** |`999 = Missing`|
| **Description** |  |
    
```{r assign_pol1, include=FALSE}

# Generally one should do:
#opal.assign.table.tibble(o, 'HAPIEE_bmi','HAPIEE_HD.hapiee_pol_w1_hd', variables=list('bmi'), missings = TRUE)
#HAPIEE_bmi <- opal.execute(o,'HAPIEE_bmi')

# Inside:
load(paste0(datafolder,"pol_w1/bmi",".RData"))
HAPIEE_bmi <- bmi
rm(bmi)
HAPIEE_bmi$bmi <- as.numeric(HAPIEE_bmi$bmi)
```

```{r local_pol1, echo=F}

vbl <- HAPIEE_bmi$bmi
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = "bmi")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(HAPIEE_bmi, aes(bmi)) + geom_histogram(stat="bin", binwidth = 0.01, fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("BMI") + ylab("Frequency") 

```

#### Harmonisation algorithm
To compute the harmonised variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo_pol1}

HAPIEE_ds_pol_w1 <- tibble(id=HAPIEE_bmi$id)
HAPIEE_ds_pol_w1$obesity <- car::recode(HAPIEE_bmi[[2]], "30:900=1; lo:30=0; NA=999")

```

#### Statistical description of the new harmonised variable
```{r descript_pol1, echo=F}

vari <- HAPIEE_ds_pol_w1$obesity
kable(Categorical_summary(vari, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(HAPIEE_ds_pol_w1, aes(x=factor(obesity))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")

```

#### Validation
```{r crosstab_pol1, echo=F}
BeforeH <- c(sum(na.omit(HAPIEE_bmi[[2]])<30),sum(na.omit(HAPIEE_bmi[[2]])>=30 & na.omit(HAPIEE_bmi[[2]])<=900), sum(na.omit(HAPIEE_bmi[[2]])==999))
AfterH <- table(HAPIEE_ds_pol_w1[[2]], useNA = "ifany")
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("<30->0",">=30->1","999") 
kable(C)
```

```{r read_pol1, include=FALSE}
rm(HAPIEE_bmi)
```


### Wave 2

Not found.


## Czech Republic

### Wave 1

#### Study-specific variable description



| **Name** |`bmi`|
|-|-|
| **Label** |`Body Mass Index calculated as weight (kg)/height(m)^2`|
| **Table name** |`hapiee_cz_w1_hd`|
| **Type of variable** |`continuous`|
| **Missings** |`999 = Missing`|
| **Description** |  |

    
```{r assign_cz1, include=FALSE}

# Generally one should do:
#opal.assign.table.tibble(o, 'HAPIEE_bmi','HAPIEE_HD.hapiee_cz_w1_hd', variables=list('bmi'), missings = TRUE)
#HAPIEE_bmi <- opal.execute(o,'HAPIEE_bmi')

# Inside:
load(paste0(datafolder,"cz_w1/bmi",".RData"))
HAPIEE_bmi <- bmi
rm(bmi)
HAPIEE_bmi$bmi <- as.numeric(HAPIEE_bmi$bmi)


```

```{r local_cz1, echo=F}

vbl <- HAPIEE_bmi$bmi
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = "bmi")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(HAPIEE_bmi, aes(bmi)) + geom_histogram(stat="bin", binwidth = 0.01, fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("BMI") + ylab("Frequency") 

```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo_cz1}
HAPIEE_ds_cz_w1 <- tibble(id=HAPIEE_bmi$id)
HAPIEE_ds_cz_w1$obesity <- car::recode(HAPIEE_bmi[[2]], "30:900=1; lo:30=0; NA=999")
HAPIEE_ds_cz_w1$obesity <- labelled(HAPIEE_ds_cz_w1[[2]], labels = c("not obese"=0, "obese"=1,  "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript_cz1, echo=F}

vari <- HAPIEE_ds_cz_w1$obesity
kable(Categorical_summary(vari, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(HAPIEE_ds_cz_w1, aes(x=factor(obesity))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")

```

#### Validation
```{r crosstab_cz1, echo=F}
BeforeH <- c(sum(na.omit(HAPIEE_bmi[[2]])<30),sum(na.omit(HAPIEE_bmi[[2]])>=30 & na.omit(HAPIEE_bmi[[2]])<=900), sum(c(sum(na.omit(HAPIEE_bmi[[2]])==999), sum(is.na(HAPIEE_bmi[[2]])) ) ))
AfterH <- table(HAPIEE_ds_cz_w1[[2]], useNA = "ifany")
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("<30->0",">=30->1","999") 
kable(C)
```


```{r read_cz, include=FALSE}
rm(HAPIEE_bmi)
```


###  Wave 2

Not found.

## Lithuania 

### Wave 1

#### Study-specific variable description



| **Name** |`bmi`|
|-|-|
| **Label** |`Body Mass Index calculated as weight (kg)/height(m)^2`|
| **Table name** |`hapiee_li_w1_hd`|
| **Type of variable** |`continuous`|
| **Missings** |`999 = Missing`|
| **Description** |  |
    
```{r assign_li1, include=FALSE}

# Generally one should do:
#opal.assign.table.tibble(o, 'HAPIEE_bmi','HAPIEE_HD.hapiee_li_w1_hd', variables=list('bmi'), missings = TRUE)
#HAPIEE_bmi <- opal.execute(o,'HAPIEE_bmi')

# Inside:
load(paste0(datafolder,"li_w1/bmi",".RData"))
HAPIEE_bmi <- bmi
rm(bmi)
HAPIEE_bmi$bmi <- as.numeric(HAPIEE_bmi$bmi)
```

```{r local_li1, echo=F}

vbl <- HAPIEE_bmi$bmi
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = "bmi")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(HAPIEE_bmi, aes(bmi)) + geom_histogram(stat="bin", binwidth = 0.01, fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("BMI") + ylab("Frequency") 

```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo_li1}
HAPIEE_ds_li_w1 <- tibble(id=HAPIEE_bmi$id)
HAPIEE_ds_li_w1$obesity <- car::recode(HAPIEE_bmi[[2]], "30:900=1; lo:30=0; NA=999")
HAPIEE_ds_li_w1$obesity <- labelled(HAPIEE_ds_li_w1[[2]], labels = c("not obese"=0, "obese"=1,  "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript_li1, echo=F}

vari <- HAPIEE_ds_li_w1$obesity
kable(Categorical_summary(vari, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(HAPIEE_ds_li_w1, aes(x=factor(obesity))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")

```

#### Validation
```{r crosstab_li1, echo=F}
BeforeH <- c(sum(na.omit(HAPIEE_bmi[[2]])<30),sum(na.omit(HAPIEE_bmi[[2]])>=30 & na.omit(HAPIEE_bmi[[2]])<=900), sum(na.omit(HAPIEE_bmi[[2]])==999))
AfterH <- table(HAPIEE_ds_li_w1[[2]], useNA = "ifany")
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("<30->0",">=30->1","999") 
kable(C)
```


```{r read_li1, include=FALSE}
rm(HAPIEE_bmi)
```





## Descriptive statistics of the harmonised variable accross populations and waves

Percentages of categories in each harmonised variable by cohort:



```{r helpfunctions, echo=F}


labelling <- function(l.hds,m.hds){
  
  # Labelling of the tibbles with categorical data and creating new tibbles with all missings recodified as NA

  for(name in names(l.hds)) {
    # In the aux_object we copy the old tibble to recodify all missing values as NA.
    aux_object <- l.hds[[name]]
    # Labelling of variables
    label(l.hds[[name]][[2]]) <- label(aux_object[[2]]) <- ds_label
    # Labelling of categories (for continues variables, only missing values)
    l.hds[[name]][[2]] <- labelled(l.hds[[name]][[2]], labels = cat_label)
    aux_object[[2]] <- car::recode(aux_object[[2]], "miss_values_vector=NA")
    # Labelling of categories (for categorical variables, only non-missing values)
    aux_object[[2]] <- labelled(aux_object[[2]], labels = cat_label[1:2])
  # Saving the recodified tibble in list m.hds
    m.hds[[name]] <- aux_object
    rm(aux_object)
  }
  return(list(l.hds,m.hds))

}

# Creation of summary tables for categorical data

summaries <- function(l.hds,m.hds,lnames){

  # Creation of columns with categories and labels
  t.hds <- frq(l.hds[[1]][2])[[1]][,c(1,2)] 
  # For each wave/population in l.hds, add the correponding values
  for (i in seq_along(l.hds)){
    t.hds[2+i] <- frq(l.hds[[i]][2])[[1]][,4] 
  }
  # Add sample size for each wave/population
  t.hds[13,] <- c("n", "sample size", sapply(l.hds,function(wave) length(wave[[1]]))
  )
  # Add wave/population names
  names(t.hds)<-c("val", "label",lnames)
  return(t.hds)
  
}

# Creation of trajectories table for each population

trajectories <- function(m.hds,vbl_name){
  
  # First wave data
  dbb <- m.hds[[1]][,c("id",vbl_name)]
  # Merge with next waves data
  for(ind in 2:length(m.hds)){
    dbb <- merge(dbb, m.hds[[ind]][,c("id",vbl_name)], by = "id", suffixes = c("", paste0(".",names(m.hds)[ind])), all = T)
  }
  names(dbb) <- c("id", names(m.hds))
  
  # Glue individual data through all waves into trajectories
  v.dbb <- dbb[,2]
  for(ind in 2:length(m.hds)){
    v.dbb <- paste(v.dbb,dbb[,ind+1],sep="")
  }
  # Trajectories and frequencies
  f.dbb <- frq(v.dbb)[[1]][,c(1,2,4)]
  return(f.dbb)
  
}

# Save data tables

savingRD <- function(l.hds,vbl_name){
  
  for(index in seq_along(l.hds)){
    assign(vbl_name,l.hds[[index]])
    save(vbl_name,list = vbl_name, file = paste0(datafolder,names(l.hds)[index],"/",vbl_name,".RData"))
  }

}

```




```{r summ, echo=F}

# All study waves and populations with abbreviated and descriptive names


hapiee.cw <- list(pol_w1 = c("pol_w1","Poland w1"), pol_w2 = c("pol_w2","Poland w2"), cz_w1 = c("cz_w1","Czech Republic w1"), cz_w2 = c("cz_w2","Czech Republic w2"), li_w1 = c("li_w1","Lithuania w1"))

# Consider only harmonised waves
l.hds <- lapply(hapiee.cw, function(wname) if(exists(paste0("HAPIEE_ds_",wname[1]))){wname = list(get(paste0("HAPIEE_ds_",wname[1])),wname[2])})
nullw <- which(sapply(l.hds, is.null))

if(length(nullw)!=0){
  l.hds <- l.hds[-nullw]
}

m.hds <- list()

lmlist <- list(lapply(l.hds,function(x)x[[1]]),m.hds)


if(length(l.hds)>0){
  # Labelling tibbles
  lmlist <- labelling(lmlist[[1]],lmlist[[2]])
  # Printing table of categories
  print(knitr::kable(summaries(lmlist[[1]],lmlist[[2]],lapply(l.hds,function(x)x[[2]])),caption=ds_label))
  # Saving tibbles
  savingRD(lmlist[[1]],vbl_name=hd_vbl)
}



```






# Quality estimation

Once BMI has been calculated using weight (Kg) and Height (m^2),
Obesity was calculated by using the BMI cutoff>=30kg/m2 following the WHO 
classification. Finally, the resulting BMI variable was coded as follows: 
BMI>30kg/m2 =1 (obese); BMI<30kg/m2 =0 (not obese) to generate two categories.


```{r closeRsession, echo=FALSE}
opal.logout(o)
```
