==> HAPIEE/6_Physical_measures/12_ah.Rmd <==
---
title: "Arterial hypertension"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---


```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_iago.r")
```

# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `ah`
* Variable label: `Arterial hypertension`
* Variable description: `Arterial hypertension. Means of all diastolic blood pressure measures.`
* Domain: `Physical Measures`
* Value type: `Categorical`
* Category coding:

**Code** | **Category Label**
-------- | ------------------
0        | no hypertension
1        | hypertension

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done 
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

```{r lab, echo=FALSE}
ds_label <- "Arterial hypertension"
ds_label_all <- "Arterial hypertension. Means of all diastolic blood pressure measures."
```


# Data process

## Poland

### Wave 1 

#### Study-specific variable description

| **Name** | `sbp`|`dbp`|
|-|-|-|
| **Label** | `Systolic blood pressure of participant at each wave` |`Diastolic blood pressure of participantat at each wave` |
| **Table name** | `sbp` |`dbp` |
| **Categories** | `continuous` |`continuous` |
| **Missings** |`991 = CAPI/interviewer error` | `991 = CAPI/interviewer error` |
| | `992 = Impute` |`992 = Impute` |
| | `993 = Disable to measure` | `993 = Disable to measure` | 
| | `994 = Not attempt/not done` | `994 = Not attempt/not done` |
| | `995 = Does not answer` | `995 = Does not answer` |
| | `996 = Not applicable` | `996 = Not applicable` |
| | `997 = Refuse` | `997 = Refuse` |
| | `998 = Do not know` |`998 = Do not know` |
| | `999 = Missing` | `999 = Missing` |
| **Description** |  |  |

```{r assign_po1, echo=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/HAPIEE/RData/pol_w1/sbp.RData")
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/HAPIEE/RData/pol_w1/dbp.RData")
HAPIEE_p1_ah <- merge(sbp, dbp, by = "id")
```

```{r local_po1, echo=FALSE}
vari <- HAPIEE_p1_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(HAPIEE_p1_ah, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(75,230)

vari <- HAPIEE_p1_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(HAPIEE_p1_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Siastolic Blood Pressure") + ylab("Frequency") + xlim(40,140)
```

#### Harmonisation algorithm
To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp>= 140 into 1`
  + `sbp< 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `NA into 999`

**R script:**

```{r harmo_po1}
HAPIEE_p1_ah$dh <- car::recode(HAPIEE_p1_ah$dbp, "0:89.99999999=0; 90:900=1")
HAPIEE_p1_ah$sh <- car::recode(HAPIEE_p1_ah$sbp, "0:139.9999999=0; 140:900=1")

HAPIEE_pol_ds_w1 <- tibble(id=HAPIEE_p1_ah$id)
HAPIEE_pol_ds_w1$ah <- NA
HAPIEE_pol_ds_w1$ah[which(HAPIEE_p1_ah$dh==0 & HAPIEE_p1_ah$sh==0)] <- 0
HAPIEE_pol_ds_w1$ah[which(HAPIEE_p1_ah$dh==1 & HAPIEE_p1_ah$sh==0)] <- 1
HAPIEE_pol_ds_w1$ah[which(HAPIEE_p1_ah$dh==0 & HAPIEE_p1_ah$sh==1)] <- 1
HAPIEE_pol_ds_w1$ah[which(HAPIEE_p1_ah$dh==1 & HAPIEE_p1_ah$sh==1)] <- 1
HAPIEE_pol_ds_w1$ah[which(HAPIEE_p1_ah$dh==999 & HAPIEE_p1_ah$sh==999)] <- 999
HAPIEE_pol_ds_w1$ah <- labelled(HAPIEE_pol_ds_w1$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_po1, echo=FALSE}
vari <- HAPIEE_pol_ds_w1$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(HAPIEE_pol_ds_w1, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation_po1, echo=FALSE}
```

#### Quality estimation
[Comments on the quality of the new harmonised variable.]

#### Import to OPAL
```{r importOPAL_po1, echo=FALSE}
#val_labels(HAPIEE_pol_ds_w1) <- NULL
#opal.assign.data(o, 'HAPIEE_pol_ds_w1', HAPIEE_pol_ds_w1)
#opal.symbol_import(o,'HAPIEE_pol_ds_w1', project='_Athlos_Harm_Dataset')
label(HAPIEE_pol_ds_w1$ah) <- ds_label_all
ah <- HAPIEE_pol_ds_w1
save(ah, file = "../RData/pol_w1/ah.RData")
rm(ah)
```





### Wave 2

There is no variable



## Czech Republic

### Wave 1 

#### Study-specific variable description

| **Name** | `sbp`|`dbp`|
|-|-|-|
| **Label** | `Systolic blood pressure of participant at each wave` |`Diastolic blood pressure of participantat at each wave` |
| **Table name** | `sbp` |`dbp` |
| **Categories** | `continuous` |`continuous` |
| **Missings** |`991 = CAPI/interviewer error` | `991 = CAPI/interviewer error` |
| | `992 = Impute` |`992 = Impute` |
| | `993 = Disable to measure` | `993 = Disable to measure` | 
| | `994 = Not attempt/not done` | `994 = Not attempt/not done` |
| | `995 = Does not answer` | `995 = Does not answer` |
| | `996 = Not applicable` | `996 = Not applicable` |
| | `997 = Refuse` | `997 = Refuse` |
| | `998 = Do not know` |`998 = Do not know` |
| | `999 = Missing` | `999 = Missing` |
| **Description** |  |  |

```{r assign_cz1, echo=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/HAPIEE/RData/cz_w1/sbp.RData")
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/HAPIEE/RData/cz_w1/dbp.RData")
HAPIEE_cz1_ah <- merge(sbp, dbp, by = "id")
```

```{r local_cz1, echo=FALSE}
vari <- HAPIEE_cz1_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(HAPIEE_cz1_ah, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(75,230)

vari <- HAPIEE_cz1_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(HAPIEE_cz1_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Siastolic Blood Pressure") + ylab("Frequency") + xlim(40,140)
```


#### Harmonisation algorithm
To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp>= 140 into 1`
  + `sbp< 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `NA into 999`

**R script:**

```{r harmo_cz1}
HAPIEE_cz1_ah$dh <- car::recode(HAPIEE_cz1_ah$dbp, "0:89.99999999=0; 90:900=1")
HAPIEE_cz1_ah$sh <- car::recode(HAPIEE_cz1_ah$sbp, "0:139.9999999=0; 140:900=1")


HAPIEE_cz_ds_w1 <- tibble(id=HAPIEE_cz1_ah$id)
HAPIEE_cz_ds_w1$ah <- NA
HAPIEE_cz_ds_w1$ah[which(HAPIEE_cz1_ah$dh==0 & HAPIEE_cz1_ah$sh==0)] <- 0
HAPIEE_cz_ds_w1$ah[which(HAPIEE_cz1_ah$dh==1 & HAPIEE_cz1_ah$sh==0)] <- 1
HAPIEE_cz_ds_w1$ah[which(HAPIEE_cz1_ah$dh==0 & HAPIEE_cz1_ah$sh==1)] <- 1
HAPIEE_cz_ds_w1$ah[which(HAPIEE_cz1_ah$dh==1 & HAPIEE_cz1_ah$sh==1)] <- 1
HAPIEE_cz_ds_w1$ah[which(HAPIEE_cz1_ah$dh==999 & HAPIEE_cz1_ah$sh==999)] <- 999
HAPIEE_cz_ds_w1$ah <- labelled(HAPIEE_cz_ds_w1$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_cz1, echo=FALSE}
vari <- HAPIEE_cz_ds_w1$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(HAPIEE_cz_ds_w1, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation_cz1, echo=FALSE}
```

#### Quality estimation
[Comments on the quality of the new harmonised variable.]

#### Import to OPAL
```{r importOPAL_cz1, echo=FALSE}
#val_labels(HAPIEE_cz_ds_w1) <- NULL
#opal.assign.data(o, 'HAPIEE_cz_ds_w1', HAPIEE_cz_ds_w1)
#opal.symbol_import(o,'HAPIEE_cz_ds_w1', project='_Athlos_Harm_Dataset')
label(HAPIEE_cz_ds_w1$ah) <- ds_label_all
ah <- HAPIEE_cz_ds_w1
save(ah, file = "../RData/cz_w1/ah.RData")
rm(ah)
```









### HAPIEE - Czech Republic - Reexamination

There is no variable.







## Lithuania

### Wave 1 

#### Study-specific variable description

| **Name** | `sbp`|`dbp`|
|-|-|-|
| **Label** | `Systolic blood pressure of participant at each wave` |`Diastolic blood pressure of participantat at each wave` |
| **Table name** | `sbp` |`dbp` |
| **Categories** | `continuous` |`continuous` |
| **Missings** |`991 = CAPI/interviewer error` | `991 = CAPI/interviewer error` |
| | `992 = Impute` |`992 = Impute` |
| | `993 = Disable to measure` | `993 = Disable to measure` | 
| | `994 = Not attempt/not done` | `994 = Not attempt/not done` |
| | `995 = Does not answer` | `995 = Does not answer` |
| | `996 = Not applicable` | `996 = Not applicable` |
| | `997 = Refuse` | `997 = Refuse` |
| | `998 = Do not know` |`998 = Do not know` |
| | `999 = Missing` | `999 = Missing` |
| **Description** |  |  |

```{r assign_li1, echo=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/HAPIEE/RData/li_w1/sbp.RData")
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/HAPIEE/RData/li_w1/dbp.RData")
HAPIEE_li_ah <- merge(sbp, dbp, by = "id")
```

```{r local_li1, echo=FALSE}
vari <- HAPIEE_li_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(HAPIEE_li_ah, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(75,230)

vari <- HAPIEE_li_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(HAPIEE_li_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Siastolic Blood Pressure") + ylab("Frequency") + xlim(40,140)
```


#### Harmonisation algorithm
To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp>= 140 into 1`
  + `sbp< 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `NA into 999`

**R script:**

```{r harmo_li1}
HAPIEE_li_ah$dh <- car::recode(HAPIEE_li_ah$dbp, "0:89.99999999=0; 90:900=1")
HAPIEE_li_ah$sh <- car::recode(HAPIEE_li_ah$sbp, "0:139.9999999=0; 140:900=1")

HAPIEE_li_ds_w1 <- tibble(id=HAPIEE_li_ah$id)
HAPIEE_li_ds_w1$ah <- NA
HAPIEE_li_ds_w1$ah[which(HAPIEE_li_ah$dh==0 & HAPIEE_li_ah$sh==0)] <- 0
HAPIEE_li_ds_w1$ah[which(HAPIEE_li_ah$dh==1 & HAPIEE_li_ah$sh==0)] <- 1
HAPIEE_li_ds_w1$ah[which(HAPIEE_li_ah$dh==0 & HAPIEE_li_ah$sh==1)] <- 1
HAPIEE_li_ds_w1$ah[which(HAPIEE_li_ah$dh==1 & HAPIEE_li_ah$sh==1)] <- 1
HAPIEE_li_ds_w1$ah[which(HAPIEE_li_ah$dh==999 & HAPIEE_li_ah$sh==999)] <- 999
HAPIEE_li_ds_w1$ah <- labelled(HAPIEE_li_ds_w1$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_li1, echo=FALSE}
vari <- HAPIEE_li_ds_w1$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(HAPIEE_li_ds_w1, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation_li1, echo=FALSE}
```

#### Quality estimation

No comments.



#### Import to OPAL
```{r importOPAL_li1, echo=FALSE}
#val_labels(HAPIEE_li_ds_w1) <- NULL
#opal.assign.data(o, 'HAPIEE_li_ds_w1', HAPIEE_li_ds_w1)
#opal.symbol_import(o,'HAPIEE_li_ds_w1', project='_Athlos_Harm_Dataset')
label(HAPIEE_li_ds_w1$ah) <- ds_label_all
ah <- HAPIEE_li_ds_w1
save(ah, file = "../RData/li_w1/ah.RData")
rm(ah)
```






```{r closeRsession, echo=FALSE}
opal.logout(o)
```
